relocate log to int 8 packers
